feat: Home, Signup and Login pages
