+++ .env.example
@@ -1,2 +1,1 @@
 NEXT_PUBLIC_ALEPH_CHANNEL=TEST
-PORT=8080
