+++ .github/milestone_template.md
@@ -25,22 +25,6 @@
 - [ ] Issues created, with the rights labels and linked to this milestone
 - [ ] Issues assigned
 
-## Monitoring
-
-### Week 1
-
-[Detail this week's report]
-
-### Week 2
-
-[Detail this week's report]\
-[If this milestone's duration is less than 2 weeks, please delete this section]
-
-### Week 3
-
-[Detail this week's report]\
-[If this milestone's duration is less than 3 weeks, please delete this section]
-
 ## Final Report
 
 ### Checklist
