+++ .gitignore
@@ -1,35 +1,42 @@
+
+# env
 .env
+.env*.local
 
+# editors
 .idea/
-
-cypress/downloads
+.vscode/
 
 # dependencies
 node_modules
 /.pnp
 .pnp.js
-.next
 
 # testing
 /coverage
+cypress/screenshots
+cypress/videos
+cypress/downloads
+
+# next.js
+/.next/
+/out/
 
 # production
 /build
 
 # misc
 .DS_Store
+*.pem
 .env.local
 .env.development.local
 .env.test.local
 .env.production.local
 
+# debug
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
-cypress/screenshots
-cypress/videos
 
 .metamask
-
-/.vscode
