+++ pages/_app.tsx
@@ -1,0 +1,67 @@
+import { useEffect, useState } from 'react';
+import { ChakraProvider, Center, Spinner, useToast } from '@chakra-ui/react';
+
+import User from '../src/lib/user';
+import Auth from '../src/lib/auth';
+
+import theme from 'theme';
+import Head from 'next/head';
+
+import '../src/theme/index.css';
+
+import UserContext from '../src/contexts/user';
+import AuthContext from '../src/contexts/auth';
+
+
+const App = ({ Component, pageProps }) => {
+	const [auth, setAuth] = useState<Auth | undefined>(undefined);
+	const [user, setUser] = useState<User | undefined>(undefined);
+	const [error, setError] = useState<Error | unknown>(undefined);
+	const toast = useToast();
+
+	useEffect(() => {
+		if (!auth && !error) {
+			try {
+				setAuth(new Auth());
+			} catch (e) {
+				setError(e);
+			}
+		}
+	}, []);
+
+	useEffect(() => {
+		if (error) {
+			toast({
+				title: 'Internal Server Error',
+				status: 'error',
+				isClosable: true,
+			});
+		}
+	}, [error]);
+
+	if (!auth) {
+		return (
+			<Center mt="160px">
+				<Spinner w="160px" />
+			</Center>
+		);
+	}
+	return (
+		<>
+			<Head>
+				<title>InterPlanerataryCloud</title>
+    	    	<meta name="description" content="Generated by our noob team" />
+    	    	<link rel="icon" href="/ipc-logo.svg" />
+			</Head>
+			<ChakraProvider theme={theme} resetCSS>
+				<AuthContext.Provider value={auth}>
+					<UserContext.Provider value={{ user: user as User, setUser }}>
+						<Component {...pageProps} />
+					</UserContext.Provider>
+				</AuthContext.Provider>
+			</ChakraProvider>
+		</>
+	);
+}
+
+export default App;
