+++ pages/signup.tsx
@@ -1,18 +1,20 @@
 import { useState } from 'react';
+import Link from 'next/link';
 
-import { Button, Link, Text, Textarea, useDisclosure, useToast, VStack } from '@chakra-ui/react';
-
-import { AuthReturnType } from '../src/lib/auth';
+import { Button, Text, Textarea, useDisclosure, useToast, VStack } from '@chakra-ui/react';
 
 import { useAuthContext } from '../src/contexts/auth';
 import { useUserContext } from '../src/contexts/user';
 
+import { AuthReturnType } from '../src/lib/auth';
+
 import Modal from '../src/components/Modal';
 import OutlineButton from '../src/components/OutlineButton';
+import AuthPage from '../src/components/AuthPage';
 
 import colors from '../src/theme/foundations/colors';
 
-const SignupView = (): JSX.Element => {
+const Signup = (): JSX.Element => {
 	const auth = useAuthContext();
 	const { setUser } = useUserContext();
 
@@ -69,44 +71,48 @@ const SignupView = (): JSX.Element => {
 	};
 
 	return (
-		<VStack spacing={{ base: '48px', md: '56px', lg: '64px' }} w="100%">
-			<Button
-				variant="inline"
-				mt="16px"
-				w="100%"
-				type="submit"
-				onClick={() => signupWithCredentials()}
-				isLoading={isLoadingCredentials}
-				id="ipc-signupView-credentials-signup-button"
-			>
-				Signup
-			</Button>
-			<VStack w="100%">
-				<Text fontSize="14px">Already an account ?</Text>
-				<Link as={RouteLink} to="/login" w="100%">
-					<OutlineButton w="100%" text="Login" id="ipc-signupView-login-button" />
-				</Link>
+		<AuthPage children={
+			<VStack spacing={{ base: '48px', md: '56px', lg: '64px' }} w="100%">
+				<Button
+					variant="inline"
+					mt="16px"
+					w="100%"
+					type="submit"
+					onClick={() => signupWithCredentials()}
+					isLoading={isLoadingCredentials}
+					id="ipc-signupView-credentials-signup-button"
+				>
+					Signup
+				</Button>
+				<VStack w="100%">
+					<Text fontSize="14px">Already an account ?</Text>
+					<Link href="/login">
+						<div style={{ width: "100%" }}>
+							<OutlineButton w="100%" text="Login" id="ipc-signupView-login-button" />
+						</div>
+					</Link>
+				</VStack>
+				<Modal
+					isOpen={isOpen}
+					onClose={closeModal}
+					title="Your Mnemonics"
+					CTA={
+						<Button variant="inline" onClick={onClick} w="100%" mb="16px" id="ipc-signupView-copy-mnemonics-button">
+							Copy my mnemonics
+						</Button>
+					}
+				>
+					<Textarea
+						value={mnemonics}
+						_focus={{ boxShadow: `0px 0px 0px 2px ${colors.red[300]}` }}
+						cursor="text"
+						readOnly
+						id="ipc-signupView-text-area"
+					/>
+				</Modal>
 			</VStack>
-			<Modal
-				isOpen={isOpen}
-				onClose={closeModal}
-				title="Your Mnemonics"
-				CTA={
-					<Button variant="inline" onClick={onClick} w="100%" mb="16px" id="ipc-signupView-copy-mnemonics-button">
-						Copy my mnemonics
-					</Button>
-				}
-			>
-				<Textarea
-					value={mnemonics}
-					_focus={{ boxShadow: `0px 0px 0px 2px ${colors.red[300]}` }}
-					cursor="text"
-					readOnly
-					id="ipc-signupView-text-area"
-				/>
-			</Modal>
-		</VStack>
+		}/>
 	);
 };
 
-export default SignupView;
+export default Signup;
